# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Chat de Atendimento - API
  description: |
    API RESTful para sistema de atendimento ao cliente via WhatsApp.
    
    ## Funcionalidades Principais
    
    - 🔐 **Autenticação JWT** com SQL Server
    - 💬 **WhatsApp Business API** (Meta Cloud)
    - 🤖 **Chatbot IA** com Google Gemini
    - ⚡ **WebSocket** para tempo real
    - 📊 **Dashboard** com métricas
    
    ## Fluxo de Autenticação
    
    1. Faça login via `POST /auth/login`
    2. Use o token retornado no header `Authorization: Bearer <token>`
    3. Token expira em 24 horas
    
  version: 2.0.0
  contact:
    name: Suporte API
    email: api@empresa.com
    url: https://empresa.com/suporte
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Ambiente de Desenvolvimento
  - url: https://api.empresa.com/api/v1
    description: Ambiente de Produção
...
paths:
...
  /users:
    get:
      tags:
        - users
      summary: Listar usuários
      description: Lista todos os usuários (somente admin)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Número da página
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Itens por página
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: role
          in: query
          description: Filtrar por role
          schema:
            type: string
            enum: [admin, atendente, supervisor, user]
        - name: is_active
          in: query
          description: Filtrar por status
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '403':
          description: Sem permissão

    post:
      tags:
        - users
      summary: Criar usuário
      description: Cria um novo usuário (somente admin)
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos
        '403':
          description: Sem permissão
  /users/{user_id}:
    get:
      tags:
        - users
      summary: Obter usuário
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuário não encontrado

    put:
      tags:
        - users
      summary: Atualizar usuário
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuário atualizado
        '404':
          description: Usuário não encontrado

    delete:
      tags:
        - users
      summary: Desativar usuário
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuário desativado
        '404':
          description: Usuário não encontrado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via /auth/login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: usuario@email.com
        nome:
          type: string
          example: João Silva
        role:
          type: string
          enum: [admin, atendente, supervisor, user]
          example: atendente
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - nome
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        nome:
          type: string
        role:
          type: string
          enum: [admin, atendente, supervisor, user]
          default: user
    UpdateUserRequest:
      type: object
      properties:
        nome:
          type: string
        role:
          type: string
        is_active:
          type: boolean
...
security:
  - bearerAuth: []